# cm.klg.service-request/specs/openapi/inbound/main.yml
---
openapi: "3.0.3"
info:
  title: "SERVICE PROVIDER MANAGEMENT REST API"
  description: "API for managing service provider"
  version: "0.0.1-SNAPSHOT"
paths:
  /service-request:
    post:
      tags:
      - "ServiceRequest"
      summary: "Create New Service Request"
      description: "Create New Service Request"
      operationId: "createNewServiceRequest"
      security:
      - bearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ServiceRequestRegister"
      responses:
        "201":
          description: "Successful operation"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreationResponse"
        "404":
          description: "SERVICE_PROVIDER_404_002"
        "409":
          description: "SERVICE-PROVIDER_409_001| SERVICE-PROVIDER_409_002"
  /my/service-request:
    get:
      tags:
      - "ServiceRequest"
      summary: "Get All My Service Requests"
      description: "Get All My Service Requests"
      operationId: "getAllMyServiceRequest"
      security:
      - bearerAuth: []
      parameters:
      - name: "status"
        in: "query"
        required: false
        schema:
          $ref: "#/components/schemas/ServiceRequestStatus"
      - name: "page"
        in: "query"
        required: false
        schema:
          type: "integer"
          minimum: "0"
      - name: "limit"
        in: "query"
        required: true
        description: "represents the number of service Request to fetch per page."
        schema:
          type: "integer"
          minimum: "10"
      responses:
        "200":
          description: "Successful operation"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServiceRequestPaginate"
components:
  schemas:
    ServiceRequestRegister:
      required:
      - "providerId"
      - "serviceTypeId"
      type: "object"
      properties:
        providerId:
          type: "string"
          format: "uuid"
        serviceTypeId:
          type: "string"
          format: "uuid"
        title:
          type: "string"
        description:
          type: "string"
        location:
          type: "string"
    ServiceRequestPaginate:
      type: "object"
      required:
      - "serviceRequest"
      - "count"
      properties:
        serviceRequest:
          type: "array"
          items:
            $ref: "#/components/schemas/ServiceRequest"
        count:
          type: "integer"
          format: "int64"
    ServiceRequest:
      required:
      - "id"
      - "userId"
      - "providerId"
      - "serviceTypeId"
      - "status"
      - "createdAt"
      type: "object"
      properties:
        id:
          type: "string"
          format: "uuid"
        userId:
          type: "string"
          format: "uuid"
        providerId:
          type: "string"
          format: "uuid"
        serviceTypeId:
          type: "string"
          format: "uuid"
        title:
          type: "string"
        description:
          type: "string"
        location:
          type: "string"
        createdAt:
          type: "string"
          format: "date-time"
        status:
          $ref: "#/components/schemas/ServiceRequestStatus"
    ServiceRequestStatus:
      type: "string"
      enum:
      - "PENDING"
      - "ACCEPTED"
      - "REJECTED"
      - "CANCELLED"
    CreationResponse:
      type: "object"
      required:
      - "newId"
      properties:
        newId:
          type: "string"
          format: "uuid"